增加 websockify

+++ get_novnc.sh
@@ -1,16 +1,20 @@
 #!/bin/bash
 
 if ! test -d noVNC; then
-	git clone --depth 1 https://github.com/novnc/noVNC
-    rm -rf .noVNC/.git
+    git clone --depth 1 https://github.com/novnc/noVNC
+    rm -rf noVNC/.git
+else
+    echo "noVNC already here"
 fi
 
-if ! test -d websockify; then
-	git clone --depth 1 https://github.com/novnc/websockify
-    rm -rf .websockify/.git
+if ! test -d noVNC/utils/websockify; then
+    git clone --depth 1 https://github.com/novnc/websockify
+    rm -rf websockify/.git
+    mv websockify noVNC/utils
+else
+    echo "websockify already here"
 fi
 
-mv websockify noVNC/utils
-cp noVNC/vnc.html noVNC/index.html
+test -e noVNC/index.html || cp noVNC/vnc.html noVNC/index.html
 sed -i -- "s/ps -p/ps -o pid | grep/g" noVNC/utils/launch.sh
